requirements.txt
streamlit
pinecone-client
sentence-transformers
torch
PyPDF2
